+++ web/app/themes/welcome.helsinki/resources/views/layouts/app.blade.php
@@ -4,6 +4,12 @@
     @include('partials.header')
   </header>
 
+  @if (is_front_page() && is_active_sidebar('front-page-banner'))
+    <div class="site__banner">
+      @php(dynamic_sidebar('front-page-banner'))
+    </div>
+  @endif
+
   <main
     class="site__content"
     @if (get_field('schema_org_is_faq_page'))